+++ api/src/controllers/WhitelistController.ts
@@ -1,9 +1,6 @@
-import {
-  Authorized, Body, Delete, Get, JsonController, Post, Param, Put, QueryParam, UseBefore,
-  HttpError, BadRequestError, UploadedFile, CurrentUser
-} from 'routing-controllers';
+import {Get, Post, Put, Delete, Authorized, Param, Body, QueryParam, CurrentUser,
+  UseBefore, UploadedFile, JsonController, BadRequestError, HttpError} from 'routing-controllers';
 import passportJwtMiddleware from '../security/passportJwtMiddleware';
-import {isNullOrUndefined} from 'util';
 import {WhitelistUser} from '../models/WhitelistUser';
 import {errorCodes} from '../config/errorCodes';
 import * as mongoose from 'mongoose';
@@ -11,27 +8,6 @@ import ObjectId = mongoose.Types.ObjectId;
 import {Course} from '../models/Course';
 import {User} from '../models/User';
 import {IWhitelistUser} from '../../../shared/models/IWhitelistUser';
-import {IUser} from '../../../shared/models/IUser';
-
-const multer = require('multer');
-import crypto = require('crypto');
-
-
-const uploadOptions = {
-  storage: multer.diskStorage({
-    destination: (req: any, file: any, cb: any) => {
-      cb(null, 'tmp/');
-    },
-    filename: (req: any, file: any, cb: any) => {
-      const extPos = file.originalname.lastIndexOf('.');
-      const ext = (extPos !== -1) ? `.${file.originalname.substr(extPos + 1).toLowerCase()}` : '';
-      crypto.pseudoRandomBytes(16, (err, raw) => {
-        cb(err, err ? undefined : `${raw.toString('hex')}${ext}`);
-      });
-    }
-  }),
-};
-
 
 @JsonController('/whitelist')
 @UseBefore(passportJwtMiddleware)
@@ -52,15 +28,11 @@ export class WitelistController {
    *        "exists": false
    *    }
    *  ]
-   * @apiParam data
-   * @apiParam currentUser
    * @apiParam whitelistToCheck
    */
   @Get('/check/:whitelist')
   @Authorized(['teacher', 'admin'])
-  async checkWhitelistForExistingStudents(@Body() data: any,
-                                          @CurrentUser() currentUser: IUser,
-                                          @Param('whitelist') whitelistToCheck: any[]) {
+  async checkWhitelistForExistingStudents(@Param('whitelist') whitelistToCheck: any[]) {
 
     return await Promise.all(
       whitelistToCheck.map(async uid => { {